requirements-dev.txt
-r requirements.txt

pytest-cov
pytest-django
model_mommy
django-test-plus